general clean-up
- Mark unused function parameters as such.
- Type corrections.
- Remove unused variables.

+++ src/wesside-ng/wesside-ng.c
@@ -1443,7 +1443,9 @@ send_fragment(struct wstate * ws, struct frag_state * fs, struct prga_info * pi)
 	*pcrc = htole32(crc);
 
 	ALLEGE(fragsize < INT_MAX - 4);
-	for (i = 0; i < (fragsize + 4) && i < (sizeof(buf) - sizeof(*wh) - 1); i++)
+	for (i = 0;
+		 i < (fragsize + 4) && (size_t) i < (sizeof(buf) - sizeof(*wh) - 1);
+		 i++)
 		body[i] ^= pi->pi_prga[i];
 
 	seq = (unsigned short *) &wh->i_seq;
